+++ packages/commonwealth/client/styles/pages/view_proposal/linked_proposals_embed.scss
@@ -2,11 +2,11 @@
 
 .LinkedProposalsEmbed {
   background-color: $neutral-100;
+  border-radius: $border-radius-corners;
   display: flex;
   flex-direction: column;
   gap: 16px;
   padding: 16px;
-  border-radius: $border-radius-corners;
 
   .treasury-embed-section {
     p {
